08.29.03细节修改

+++ src/com/zzxx/music/dao/SongDao.java
@@ -36,7 +36,7 @@ public interface SongDao {
 
     public List<Search> selectAllSource(String name);
     
-    Song getSongBySongId(int songid);
+    Song getSongBySongId(long songid);
     
     
 }
